refactor: reorganize files & expand doc
