+++ rccheck/src/tests/ed25519_certgen_tests.rs
